Revamp The Login System

+++ app/Http/Controllers/ReportController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Services\Utility;
 use App\Models\HeadReport;
+use Illuminate\Support\Facades\Gate;
 
 class ReportController extends Controller
 {
@@ -85,6 +86,9 @@ class ReportController extends Controller
      */
     public function edit($maintenance_type, $id)
     {
+        $headReport = HeadReport::findOrFail($id);
+        $this->authorize('update', $headReport);
+        
         return view('expert.report.edit', compact('id', 'maintenance_type'));
     }
 
@@ -96,6 +100,9 @@ class ReportController extends Controller
      */
     public function destroy($maintenance_type, $id)
     {
+        $headReport = HeadReport::findOrFail($id);
+        $this->authorize('update', $headReport);
+
         HeadReport::destroy($id);
         return redirect()->route('report.index', $maintenance_type)->with('status_delete', 'Data Dihapus');
     }
